feat: basic dynamic
